feat: add CRON job to prewarm relay test cache
