+++ .env.example
@@ -13,3 +13,6 @@ RPC_ENDPOINTS_JSON=[{"name":"gateway-a","rpcUrl":"https://path.portal.gateway-a.
 # RELAY_TEST_ENDPOINT_CONCURRENCY=3
 # RELAY_TEST_ENDPOINT_TIMEOUT_MS=4000
 # RELAY_TEST_HEALTH_TIMEOUT_MS=2500
+
+# Optional: protects /api/relay-test/prewarm (used by Vercel Cron)
+# CRON_SECRET=replace-with-a-long-random-string
