+++ src/app/api/relay-test/route.ts
@@ -1,129 +1,29 @@
 import { NextResponse } from "next/server";
-import servicesJson from "@/data/services.json";
 import {
-  getCoverageByServiceId,
-  mapWithConcurrency,
-  performRelayTestForChain,
-} from "@/utils/relay-test";
-import { Chain } from "@/utils/types";
-
-const RESPONSE_TTL_MS = parsePositiveInt(
-  process.env.RELAY_TEST_RESPONSE_TTL_MS,
-  5 * 60 * 1000,
-);
-const CHAIN_TEST_CONCURRENCY = parsePositiveInt(
-  process.env.RELAY_TEST_CHAIN_CONCURRENCY,
-  8,
-);
-
-type RelayApiRow = {
-  label: string;
-  serviceId: string;
-  chain: string;
-  type: string;
-  blockNumber: string | undefined;
-  status: "success" | "error";
-  latency: number;
-};
-
-type RelayResponseCache = {
-  result: RelayApiRow[];
-  expiresAt: number;
-};
-
-let relayResponseCache: RelayResponseCache | null = null;
-let relayResponseRefreshPromise: Promise<RelayApiRow[]> | null = null;
-
-function parsePositiveInt(value: string | undefined, fallback: number): number {
-  if (!value) {
-    return fallback;
-  }
-  const parsed = Number.parseInt(value, 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : fallback;
-}
-
-function cacheHeaders(): Record<string, string> {
-  const ttlSeconds = Math.max(1, Math.floor(RESPONSE_TTL_MS / 1000));
-  return {
-    "Cache-Control": `public, max-age=0, s-maxage=${ttlSeconds}, stale-while-revalidate=60`,
-  };
-}
-
-async function computeRelayApiRows(): Promise<RelayApiRow[]> {
-  const coverageByServiceId = await getCoverageByServiceId();
-  const chains = (servicesJson.services as Chain[]).filter((chain) =>
-    coverageByServiceId.has(chain.serviceId),
-  );
-
-  return mapWithConcurrency(chains, CHAIN_TEST_CONCURRENCY, async (chain) => {
-    const result = await performRelayTestForChain(
-      chain,
-      coverageByServiceId.get(chain.serviceId) || [],
-    );
-
-    return {
-      label: chain.label,
-      serviceId: chain.serviceId,
-      chain: chain.name,
-      type: chain.type,
-      ...result,
-      blockNumber: result.blockNumber?.toString(),
-    };
-  });
-}
-
-async function refreshRelayResponseCache(): Promise<RelayApiRow[]> {
-  if (!relayResponseRefreshPromise) {
-    relayResponseRefreshPromise = computeRelayApiRows()
-      .then((result) => {
-        relayResponseCache = {
-          result,
-          expiresAt: Date.now() + RESPONSE_TTL_MS,
-        };
-        return result;
-      })
-      .finally(() => {
-        relayResponseRefreshPromise = null;
-      });
-  }
-
-  return relayResponseRefreshPromise;
-}
-
-async function getRelayApiRowsCached(): Promise<RelayApiRow[]> {
-  const now = Date.now();
-  if (relayResponseCache && now < relayResponseCache.expiresAt) {
-    return relayResponseCache.result;
-  }
-
-  if (relayResponseCache) {
-    void refreshRelayResponseCache().catch((error) => {
-      console.warn(
-        "[relay-test] Failed to refresh response cache; serving stale data:",
-        (error as Error).message,
-      );
-    });
-    return relayResponseCache.result;
-  }
-
-  return refreshRelayResponseCache();
-}
+  getRelayApiRowsCached,
+  getRelayApiRowsSnapshot,
+  relayTestCacheHeaders,
+} from "@/utils/relay-test-cache";
 
 export async function GET() {
   try {
     const responses = await getRelayApiRowsCached();
-    return NextResponse.json({
-      success: true,
-      result: responses,
-    }, { headers: cacheHeaders() });
+    return NextResponse.json(
+      {
+        success: true,
+        result: responses,
+      },
+      { headers: relayTestCacheHeaders() },
+    );
   } catch (error) {
-    if (relayResponseCache) {
+    const cached = getRelayApiRowsSnapshot();
+    if (cached) {
       return NextResponse.json(
         {
           success: true,
-          result: relayResponseCache.result,
+          result: cached,
         },
-        { headers: cacheHeaders() },
+        { headers: relayTestCacheHeaders() },
       );
     }
