+++ vercel.json
@@ -1,0 +1,8 @@
+{
+  "crons": [
+    {
+      "path": "/api/relay-test/prewarm",
+      "schedule": "*/5 * * * *"
+    }
+  ]
+}
